Text Package: *m commit_line.py - Commit Text Organizer Package *m commit_line_prefixes.py - Commit Text Organizer Package *m commit_text_group.py - Commit Text Organizer Package *m commit_text_organizer.py - Commit Text Organizer Package
